+++ NEWS.md
@@ -1,4 +1,4 @@
-# arules 1.6-5.1 (xx/xx/2020)
+# arules 1.6-6 (05/14/2020)
 
 ## New Features
 * added interestMeasure rhsSupport.